pyproject.toml
[tool.poetry]
name = "iwrok_test"
version = "0.1.0"
description = "A test by IWork"
authors = ["Mohamed Nesredin <m.n.kaizen@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "example.com"
documentation = ""
repository = "https://github.com/Mohamed-Kaizen/IWork-test/"
keywords = ["iwrok_test"]

[tool.poetry.dependencies]
python = "3.8.6"
argon2-cffi = "^20.1"
confusable_homoglyphs = "^3.2.0"
dj-database-url = "^0.5"
dj-rest-auth = {extras = ["with_social"], version = "^1.1"}
django = "^3.1"
djangorestframework = "^3.12"
djangorestframework-simplejwt = "^4.4.0"
django-axes = "^5.7"
django-cors-headers = "^3.4"
django-user-g11n = "^0.4"
drf-spectacular = "^0.9"
gunicorn = "^20.0"
psycopg2-binary = "^2.8"
python-decouple = "^3.3"
pwned-passwords-django = "^1.4"
whitenoise = {extras = ["brotli"], version = "^5.1"}

[tool.poetry.dev-dependencies]
black = "^20.8b1"
coverage = {extras = ["toml"], version = "^5.2"}
django-coverage-plugin = "^1.8"
django-debug-toolbar = "^3.1"
darglint = "^1.5"
flake8 = "^3.8"
flake8-annotations = "^2.4"
flake8-bandit = "^2.1"
flake8-black = "^0.2"
flake8-bugbear = "^20.1"
flake8-docstrings = "^1.5"
flake8-import-order = "^0.18"
isort = "^5.5.4"
mypy = "^0.782"
mkdocs = "^1.1"
mkdocs-material = "^6.0"
mkdocstrings = "^0.13"
mkdocs-minify-plugin = "^0.3"
mkdocs-git-revision-date-localized-plugin = "^0.7"
pytest = "^6.1"
pytest-cov = "^2.10"
pytest-django = "^3.10"
pytest-django-queries = "^1.1"
radon = "^4.3"
safety = "^1.9.0"
xdoctest = "^0.15"
pytest-dotenv = "^0.5.2"

[tool.coverage.run]
branch = true
source = ["iwrok_test"]
plugins = "django_coverage_plugin"

[tool.coverage.report]
show_missing = true
fail_under = 84

[tool.isort]
profile="black"

[tool.pytest.ini_options]
minversion = "6.0"
DJANGO_SETTINGS_MODULE="iwrok_test.settings"
env_files = ".test.env"
python_files = "tests.py test_*.py *_tests.py"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
